storage: index state decoders should throw on failure
The encoders for the deprecated version originally returned std::nullopt
to indicate failure. However, serde primarily uses a uniform approach
based around `serde::serde_exception` to transmit failures when encoding
or decoding. In order to help unify the serde and the compatibility
layer we switch the compat layer to throw as well and translate back to
a std::nullopt in the caller.

Signed-off-by: Noah Watkins <noah@vectorized.io>