test: throw on failed cmd retries
The intention of kcl::_cmd is to throw when retryies of a command fails.
However, the formulation allows execution to return None is some cases
(e.g. if zero retries was passed in). In any case, returning None is
problematic because it gives the return type of Optional[X] which means
all users need to explicitly check for the error condition that they
already assume won't occur. By throwing there are better error messages
(e.g. avoiding deferencing None) and no explicit checking is needed to
make the type checker happy.

Signed-off-by: Noah Watkins <noah@vectorized.io>