raft: allow blocking of new raft leadership
Introduces [un]block_new_leadership interface to consensus which allows
a raft group to be placed into (and removed from) a state in which new
leadership is prevented. This is used when draining leadership from a
node where preventing new leadership allows progress in draining
leadership.

A node priority override is used to prevent a node from ever calling for
a vote for itself (ie becoming a candidate). This uses the voter
priority mechanism.

Signed-off-by: Noah Watkins <noah@redpanda.com>