redpanda: add node draining manager
The node draining manager is responsible for draining (and undraining)
leadership from a node. A fiber on each core reacts to these requested
states and tracks the status of the process.

Draining proceeds by blocking all new leadership, and then proceeds to
transfer leadership for all partitions away from the current node.
Transfers are done in batches and statistics are maintained which
can be retrieves via the status api.

Each batch for leadership transfer is randomly selected. This has a
natural affect of backing off transfer for partitions that previously
encountered errors, as well as making progress across the total set of
partitions even if some are stuck. This is useful for policies in which
we want to drain within a certain time bound, but will ultimately force
a node to shutdown.

Signed-off-by: Noah Watkins <noah@redpanda.com>