cluster: block leadership for partitions
Adds a mechanism into the partition manager that allows leadership to be
blocked for both existing and newly constructed raft groups. Blocking
for new groups is useful because while a node is draining leadership
partitions may be created (e.g. the controller is still processing
requests as a follower). In these cases we can avoid any initial vote
dispatch.

Signed-off-by: Noah Watkins <noah@redpanda.com>