raft: reject leadership transfer for blocked groups
Explicit leadership transfer normally overrides any voter priority
setting for a node. This is fine, but we do not want this override to
take precedence when a group has leadership explicitly blocked. When
handling a timeout_now request we check if leadership is blocked
and return an error without dispatching a vote.

Signed-off-by: Noah Watkins <noah@redpanda.com>