kafka: fix unused lambda capture warning
The code generator could figure this out and not capture the version,
but this would require some changes that would allow some forward
looking analysis in the ast before we printed out this part of the
template. and generally nearly all generated code will eventually want
to peek at the version as the apis evolve.

Signed-off-by: Noah Watkins <noah@redpanda.com>