rpc: skip payload when handling method-not-found error
When a service method was not found a method-not-found error was
returned to the client but the payload from the client was not skipped
over by the server.

This means that for all pratical purposes the connection is corrupted
after the server handles a method not found error because it will
immediate attempt to decode what is on the wire as the header for a new
rpc but which in reality will likely be some rpc message payload.

this commit skips over the payload in this case allowing the connection to
remain usage.

Signed-off-by: Noah Watkins <noah@redpanda.com>