rpc: expand tests for method not found error
The issue in which a connection that handles a missing method becomes
unusable was not discovered because the testing in place did not attempt
to use the connection following the missing method error.

This commit expands testing to include a mixture of missing methods and
normal methods that should succeed, and confirms that each individual
request on the same connection behaves as expected.

Signed-off-by: Noah Watkins <noah@redpanda.com>