rpc: expose server max version supported to client
A server that returns an unsupported_version error should also return
the max version that it supports so that receiver can adapt to the
situation.

This commit exposes the version through the low-level transport
interface for typed message sending, and adds a test to make sure the
server retains this behavior.

Signed-off-by: Noah Watkins <noah@redpanda.com>