rpc: use dynamic version for client transport
The version of the transport represents a common version between the
client and server. When a client/server agree to mutually support a
higher version the version may be upgraded which will affect subsequent
requests.

For testing purposes, a client transport operating at v0 will
never be upgraded. This is used to simulate an old v0 client that
doesn't support any version other than v0.

Signed-off-by: Noah Watkins <noah@redpanda.com>