rpc: use variable client transport version
The version of the client transport indicates which level of encoding
(adl, serde) a client should use when sending messages to a server.
Versions v0,1 corresponds to adl<> and v2 correponds to serde.

When at version v1 a client may upgrade the connection to v2 if it
discovers that its peer supports v2. For testing, v0 connections are
never upgraded.

This patch preserves existing behavior by initializing a transport with
versions v0.

Signed-off-by: Noah Watkins <noah@redpanda.com>