rpc: add type trait to exempt types from adl<> encoding
the transition from adl to serde encoding in rpc requires a period of time
where both encodings are supported for all message types. however, we do not
want to extend this requirement to brand new messages / services. we enlist
the help of the type system to enforce these rules, but it means we need a
way to opt-out of adl support on a case-by-case basis for new message types.

the `rpc_adl_exempt` type trait helper can be used to opt-out a type T from
adl support. a type is marked exempt by defining the type T::rpc_adl_exempt.
the typedef may be defined as any type such as std::{void_t, true_type}.

Example:

    struct exempt_msg {
        using rpc_adl_exempt = std::true_type;
        ...
    };

Signed-off-by: Noah Watkins <noah@redpanda.com>