rpc: add v0 service creation in rpc test fixture
This allows the services used in rpc testing to be started in a v0 mode
for emulating different scenarios such as a new client connected to an
old server.

This also adds code generation for serde-only types in encode/decode
which unconditionally assert(false). This is because we combine all the
scenarios we want to test into a single service, and if we didn't do
this, then the service would not compile because it also contains
methods that are serde-only. This is disallowed at runtime--tests should
never be written that cause a service to handle a serde message in a v0
emulation scenario. The alternative would be a lot more bifurcation of
test services or some sort of sfinae magic.

This does not represent an issue for non-test cases because the default
message handler generates correct code for all scenarios.

Signed-off-by: Noah Watkins <noah@redpanda.com>