rpc: transport request adaptive encoding selection
When a message is sent the active transport version determines if it is
sent with adl (v0,1) or serde encoding (v2). Serde-only messages require
a version >= v2. Messages with adl-only support are always sent at v0.
Support for adl-only messages is a temporary solution to allow changes
to the RPC transport and servers to be worked on prior to all types
being supported by serde. Once all types have serde support the
allowance at the type-system level should be removed.  Sending a message
does not upgrade the transport. That is done when handling the reply.

This commit also changes the not_supported_version test to use the
echo_adl_serde variant rather than the original version whose types
cause it to be treated as adl-only. Because adl-only messages are sent
at v0 and the test specifically requires sending messages at an
arbitrarily higher value to trigger the error, a type was needed that
supports a dynamic version range. When encoding adl/serde supported
types the version is passed through.

Signed-off-by: Noah Watkins <noah@redpanda.com>