cluster: add missing default equality operators
The envelope type has a default equality operator which means that
comparing types that inherit will still compare to each other, but use
the compartor in the base class (envelope). this is largely broken
because forgetting the operator== in the derived class means comparison
always succeeds. we are going to remove the operator in the base class
but these are still needed.

Signed-off-by: Noah Watkins <noah@redpanda.com>