raft: support default ctor for append_entries_request
Serde requires that types are default constructable. Since
append_entries_request contains a record_batch_reader it is
not default construtable. Until serde is improved in this
area and provides interfaces/customization points for dealing
with such types we need to make this request default ctor.

To do this we wrap the reader in std::optional and provide
an accessor to the underlying value. The only case then
in which the optional is not set is when default constructor
is invoked by serde.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>