serde: better async handling for move-only types
1. reading moves the result into the returned ready future.
2. writing takes a value rather than a const-ref

For 2 the motivation is that serialization may require non-const actions
like reading from a record batch reader (e.g. serializing an append
entries request). In this case the reference cannot be const. However,
if we make it non-const then we can't pass temporaries and otherwise the
call sites are more confusing.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>