compat: add json helper
The compat framework requires reading/writing json for data types that
are involved in the compat checking.

There is some organization of json in the tree in the form of the
rjson_serialize interface in json/json.h but virtually no interfaces for
reading json.

Since the compat framework will be the source of creation of many json
encoder/decoders in the tree and we don't yet have a coherent strategy
for adding json support we'll concentrate that support here in the
compat framework.

The body of encoder and decoder work can largely be transplanted into
any future interfaces we converge on.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>