pp: initialize external mitigator earlier in ctor
This is not a bug fix, but rather a hygeine improvement.  Both producer
and the update metadata callback can invoke the migitigator. It is true
that this won't happen in the constructor but it's nice to have the
initialization happen in a order that doesn't require investigating if
the invokation of these uses is possible.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>