pp/server: remove dangling reference
The path description reference stored in the probe is a reference to a
stack allocated value in route_t that goes out of scope when
server::route(server::route_t r) returns. The path is only used in the
constructor, so this doesn't fix a bug, but it does eliminate the
dangling reference which would be a use-after-free bug if future code
touched the path reference.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>