kafka: disable aggregate init for offset metadata
Aggregate initialization allows callers to initialize less than the
number of members in a struct. This means we can do stuff like

  struct t { a, b };
  t v{.b = ... };

and v.a will have a default initialized value. this is problematic when
we'd like to add a new field to the type and then make sure that all
call sites are updated. it's simply hard to find them.

this commit disables aggregate initialization by adding a constructor,
forcing callers to explicitly adhere to the constructor requirements.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>