storage: move statvfs call into storage::node
This commit moves the file system stat call into storage::node where new
interfaces will be added for overriding it for test purposes via the
admin interface. This is also the first step in combining storage::node
and local_monitor in a single service, which have some split of
responsibility but end up working together in a difficult to understand
way.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>