rmgmt: block kafka writes when log max size is exceeded
When the configured max size for log storage is exceeded writes at the
kafka layer are blocked.

The bulk of new data into the system will arrive via the kafka produce
api. Notably the max size is not enforced for all incoming data to
the log layer. This might be something to consider in the future, but it
is unclear what the implications are. For example, it may be the case
that configuration batches in the controller log need to be replicated,
or that raft configuration batches need to land in order to resolve the
situation in which the max size has been exceeded.

Other Kafka API that are generally expected to result in a small amount
of data creation (e.g. consumer groups) are not blocked so that other
APIs like fetching would continue to function.

For consideration, a flexible reservation API like the one used in the
cloud storage cache would allow this distinction to be made by callers.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>