storage: add segment categorization for space management
Adds categorization of partition segments for space management. The
categories of segments are co-designed with an eviction policy that is
introduced later on in the commit series.

Roughly the categories define a priority of data to make available for
removal when the system is in a low disk situation. For example, data
that is allowed to grow past local retention rules can be considered
"best-effort" space usage, and is the first to be reclaimed. The
categorizes proceed all the way down until the active segment.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>