storage: factor out retention_offset with timestamp adjustments
The disk space management code performs a trial run with retention
settings using the retention_offsets() function. however, applying
adjustments to the future timestamps is an optional config that can be
turned on, but wasn't easily integrated into the trail run because it
was done in the normal path where retention was actually applied.

this commit factors out these two operations from the normal retention
path so that it can be reused by space management.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>