config: add generic disk reservation config option
Reserve space on underlying storage device that Redpanda will avoid
using. Additional division of the disk will be made on the space
remaining after the reservation is applied.

The primary purpose is two fold. First, it is used for both wiggle
room--our target reservations on cache and data are not precise and is
expected that they are overshot. The second reason is performance. As a
disk reaches capacity performance for some workloads will degrade.
Running at higher level of capacities should be an opt-in choice.

Default reservation is zero on an upgraded cluster.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>