datalake: add scheduler request for immediate shutdown
This commit introduces a new interface:

   scheduler::request_immediate_finish(vector<translator>)

which can be used by an entity external to the scheduler for requesting
that a set of translators be promptly stopped and finished. The primary
use case for this interface are global resource schedulers (e.g. disk
space) which want to inject priority scheduling decisions.

Translators requested for immediate finish are scheduled into the
running state, and arranged to immediately follow the code path for
finishing a translator without building a reader or poking the
translator context.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>