serde: parquet row writer should return total buffered
Prior to this commit writer::write_row would return the total bytes
buffered in memory for the current page, rather than the total number of
bytes pending to be flushed.

The later definition aligns the return value with the behavior of
writer::stat() and is more natural too: users of writer are likely
uninterested in how much data is in the current "page".

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>