cloud_topics: introduce manager
Adds scaffolding and wires up a cloud topics manager.

The purpose of the cloud topics manager is to manage state and
operations that are cluster-wide in nature. For example, L0 GC is
inherently a global operation over all cloud topics NTPs. Other examples
include operations on L1 domains.

Currently the manager will be used for L0 GC which in its current form
is a stateless operation, and doesn't need to actually store persistent
state. So we simplify the design by attaching the manager to the leader
of partition 0 of the cloud topics L1 domain topic. At which point we
need to store persistent state we'll likely want to instantiate a
dedicated partition for the manager.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>